Move downloading files logic to util
